% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MDI.R
\name{MSDI}
\alias{MSDI}
\title{Compute the Multivariate Standardized Drought Index(MSDI)}
\usage{
MSDI(X, Y, ts = 6)
}
\arguments{
\item{X}{is The vector of a monthly hydro-climatic variable of n years.}

\item{Y}{is The vector of a monthly hydro-climatic variable of n years.}

\item{ts}{is the accumulated time scale.}
}
\value{
The MSDI of different time scales from the marginal probability (or percentile)
}
\description{
Based on the vector of a monthly hydro-climatic variable,
the MSDI is computed based on the marginal (or univariate) probability (or percentile) from the function SDI.
}
\examples{
X=runif(120, min = 0, max = 100)
Y=runif(120, min = 0, max = 100)
fit<-MSDI(X,Y,ts=6) # Compute the 6 month drought index
fit$ProbEmp2 #Get the empirival drought index
}

